+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Changed
+ * Stop post-processing timestamps.
+
 ## 1.0.2 - 2022-12-21
 ### Added
  * Support for Python 3.11.